+++ src/test/java/Generic/GenericTest.java
@@ -8,19 +8,19 @@ import org.junit.jupiter.api.Assertions;
     @Test
     public void GetMaximum() {
         Generic max= new Generic();
-        int value = (int) max.Maximum(30,20,10);
-        Assertions.assertEquals(30, value);
+        Double value = (Double) max.Maximum(30.0,20.0,10.0);
+        Assertions.assertEquals(30.0, value);
     }
     @Test
         public void GetMaximum() {
             Generic max= new Generic();
-            int value = (int) max.Maximum(20,30,10);
-            Assertions.assertEquals(30, value);
+        Double value = (Double) max.Maximum(20.0,30.0,10.0);
+            Assertions.assertEquals(30.0, value);
         }
         @Test
         public void GetMaximum() {
             Generic max= new Generic();
-            int value = (int) max.Maximum(10,20,30);
+            Double value = (Double) max.Maximum(10.0,20.0,30.0);
             Assertions.assertEquals(30, value);
         }
 }
